+++ src/lib/target-generators/index.ts
@@ -1,5 +1,4 @@
 import * as node from "./node";
-import * as test from "./test";
 import * as webpack from "./webpack";
 
-export {node, test, webpack};
+export {node, webpack};