+++ src/lib/target-generators/node.ts
@@ -147,7 +147,7 @@ export function generateTarget(gulp: Gulp, project: Project, target: NodeTarget)
   });
 
   // <targetName>:clean
-  let cleanOptions: clean.Options;
+  let cleanOptions: clean.CleanOptions;
   if (target.clean !== undefined) {
     cleanOptions = {
       base: locations.rootDir,