+++ src/lib/task-generators/index.ts
@@ -2,9 +2,9 @@ import * as buildTypescript from "./build-typescript";
 import * as buildWebpack from "./build-webpack";
 import * as clean from "./clean";
 import * as copy from "./copy";
+import * as testNode from "./mocha";
 import * as pug from "./pug";
 import * as sass from "./sass";
-import * as testNode from "./test-node";
 import * as tsconfigJson from "./tsconfig-json";
 
 export {buildTypescript, buildWebpack, clean, copy, pug, sass, testNode, tsconfigJson};